+++ assets/src/scripts/components/hydrograph/index.js
@@ -17,6 +17,7 @@ import {Actions as floodDataActions} from '../../store/flood-inundation';
 import {renderTimeSeriesUrlParams} from '../../url-params';
 
 import {drawDateRangeControls} from './date-controls';
+import {drawDataTable} from './data-table';
 import {lineSegmentsByParmCdSelector} from './drawing-data';
 import {drawGraphBrush} from './graph-brush';
 import {drawGraphControls} from './graph-controls';
@@ -142,7 +143,7 @@ export const attachToNode = function (store,
                 .classed('ts-legend-controls-container', true)
                 .call(drawTimeSeriesLegend, store);
 
-            // Add UI interactive elements and the provisional data alert.
+            // Add UI interactive elements, data table  and the provisional data alert.
             if (!showOnlyGraph) {
                 nodeElem
                     .call(drawMethodPicker, store)
@@ -151,7 +152,11 @@ export const attachToNode = function (store,
                 nodeElem.select('.ts-legend-controls-container')
 
                     .call(drawGraphControls, store);
-
+                nodeElem.select('#iv-data-table-container')
+                    .call(drawDataTable, store);
+                nodeElem.select('.provisional-data-alert')
+                    .attr('hidden', null);
+                //TODO: Find out why putting this before drawDataTable causes the tests to not work correctly
                 nodeElem.select('.select-time-series-container')
                     .call(link(store, plotSeriesSelectTable, createStructuredSelector({
                         siteno: () => siteno,
@@ -159,8 +164,7 @@ export const attachToNode = function (store,
                         lineSegmentsByParmCd: lineSegmentsByParmCdSelector('current', 'P7D'),
                         timeSeriesScalesByParmCd: timeSeriesScalesByParmCdSelector('current', 'P7D', SPARK_LINE_DIM)
                     }), store));
-                nodeElem.select('.provisional-data-alert')
-                    .attr('hidden', null);
+
 
                 renderTimeSeriesUrlParams(store);
             }